request js 추가 - 파이썬 스크립트 실행시킬 controller 이다.

+++ models/KEYWORD.js
@@ -30,24 +30,22 @@ https://1-day-1-coding.tistory.com/51
 CURD MODEL 만들기
 */
 
-MakeWhereWord = (byKEYWORD) => {
-query = " WHERE "
-let cnt = 0;
-for(const id in byKEYWORD ){
-    
+const MakeWhereWord = (byKEYWORD) => {
+    query = " WHERE "
+    let cnt = 0;
+    for(const id in byKEYWORD ){
+        
 
-    if(byKEYWORD[id] != null){
-        if(cnt == 0 ) { // 첫번째는 AND 를 안붙임.
-            cnt += 1;
-            query += '`'+id + '`' + " = " +'"'+ byKEYWORD[id]+ '"';
-        }else{
-            query += " AND "+ '`'+id + '`' + " = " + '"'+byKEYWORD[id] + '"';
+        if(byKEYWORD[id] != null){
+            if(cnt == 0 ) { // 첫번째는 AND 를 안붙임.
+                cnt += 1;
+                query += '`'+id + '`' + " = " +'"'+ byKEYWORD[id]+ '"';
+            }else{
+                query += " AND "+ '`'+id + '`' + " = " + '"'+byKEYWORD[id] + '"';
+            }
         }
         
-
     }
-        
-}
 return query
 }
 KEYWORD.create  = (newKEYWORD, result) =>{
@@ -94,25 +92,24 @@ return 200;
 }
 
 
-KEYWORD.remove = (byKEYWORD, result) =>  {
+KEYWORD.removeByKeyword= (byKeyword, result) =>  {
 
-let query = "DELETE FROM KEYWORD "
-query = query + MakeWhereWord(byKEYWORD)
+    let query = "DELETE FROM KEYWORD where userid = " + byKeyword.userid + " and keyword = '" + byKeyword.keyword + "'"
 
-sql.query(query, (err, res) => {
-        if(err){
-            console.log("error : ", err);
-            result(err, null);
-            return;
-        }
+    sql.query(query, (err, res) => {
+            if(err){
+                console.log("error : ", err);
+                result(err, null);
+                return;
+            }
 
-        if(res.affectedRows == 0){
-            result({kind:"not_found"}, null)
-            return;
-        }
+            if(res.affectedRows == 0){
+                result({kind:"not_found"}, null)
+                return;
+            }
 
-        result(null, res);
-})
+            result(null, res);
+    })
 }
